old/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
qdrant-client>=1.4.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Gemini API for enhanced answer quality
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Quantum computing
pennylane>=0.32.0
qiskit>=0.44.0
qiskit-ibm-runtime>=0.13.0

# Evaluation metrics
nltk>=3.8
rouge-score>=0.1.2
bert-score>=0.3.13

# Data processing
pandas>=2.0.0
datasets>=2.14.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
jsonlines>=3.1.0
click>=8.1.0
